Finished documenting the source code for the main scripts
The main scripts (Trainer, Scorer and Cutter) have all functions documents with
doctrings, the name variables are sane and the layout of the code is more
rational.

mirscanModule.py is not ready, pending evaluation of the hairpin  criteria
feature objects. Functions below the file parsing methods are perhaps best
placed in a separate file since they are only used for the criteria rule files.