Fixed bug in the example criteria files (.py)
In the criteria example files provided, the mirscan function loops through the
list of candidates using 'for c in candidates:'. However in the past it was a
'for n in len(candidates): c = candidates[n]'. There was another usage for the
variable n in calling the appropriate member of the starts list. This change
caused an error where n was unavailable inside the function and the code
resorted to a global variable with the same name.