New format for criteria file (.py)
Criteria files (.py) define Feature and DataItem classes which are imported
into the Criteria class. DataItems are computed for each candidate and provide
temporary data (cand_data) on each of them. Features compute scores according
to the functions they define.

mirscanCriteria.py no longer needs to be altered when extra temporary data is
required for a Feature given the possibility to define extra DataItems.
mirscanCriteria.py is now self-contained, it only contains the functions
necessary for stating the sequence and structure (and alignement).

examples/criteria.Ruby_etal.py is the translation of examples/oneSeq.rubyEtAl.py
into the new format. It is not yet complete and examples/twoSeq.rubyEtAl.py
is to be translated in the future.